requirements.txt
asgiref==3.3.4
cycler==0.10.0
Django==3.2.4
django-phonenumber-field==5.2.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
phonenumbers==8.12.24
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
